cloudtests: extract CloudtestApplicationBase
